+++ NEWS.md
@@ -1,6 +1,9 @@
 
+# 2018-10-24
+* add to practice 13b sample (WebKit)
+
 # 2018-10-13
-* add to practice 12b sample
+* add to practice 12b sample (Gtk)
 
 # 2018-09-29
 * convert @pyimport to pyimport() (06b/07b)